improve error tooltip styles
